Updates to optional horizontal scrollbar.
- move SizeHintWidget out to separate files
- remove redundant logic in view.cpp
- remove sizehint from dialogue subview
- add tooltip hints in user preferences
