+++ apps/opencs/view/doc/sizehint.cpp
@@ -1,0 +1,17 @@
+#include "sizehint.hpp"
+
+CSVDoc::SizeHintWidget::SizeHintWidget(QWidget *parent) : QWidget(parent)
+{}
+
+CSVDoc::SizeHintWidget::~SizeHintWidget()
+{}
+
+QSize CSVDoc::SizeHintWidget::sizeHint() const
+{
+    return mSize;
+}
+
+void CSVDoc::SizeHintWidget::setSizeHint(const QSize &size)
+{
+    mSize = size;
+}
