+++ apps/opencs/view/doc/subview.cpp
@@ -43,19 +43,3 @@ void CSVDoc::SubView::closeRequest()
 {
     emit closeRequest (this);
 }
-
-CSVDoc::SizeHintWidget::SizeHintWidget(QWidget *parent) : QWidget(parent)
-{}
-
-CSVDoc::SizeHintWidget::~SizeHintWidget()
-{}
-
-QSize CSVDoc::SizeHintWidget::sizeHint() const
-{
-    return mSize;
-}
-
-void CSVDoc::SizeHintWidget::setSizeHint(const QSize &size)
-{
-    mSize = size;
-}
