+++ apps/opencs/view/doc/subview.hpp
@@ -8,7 +8,6 @@
 #include "subviewfactory.hpp"
 
 #include <QDockWidget>
-#include <QSize>
 
 class QUndoStack;
 
@@ -70,18 +69,6 @@ namespace CSVDoc
         protected slots:
 
             void closeRequest();
-    };
-
-    class SizeHintWidget : public QWidget
-    {
-            QSize mSize;
-
-        public:
-            SizeHintWidget(QWidget *parent = 0);
-            ~SizeHintWidget();
-
-            virtual QSize sizeHint() const;
-            void setSizeHint(const QSize &size);
     };
 }
